refactored streams
